[VT]: Added a base class for VT object pools
Added a base class for VT object pools so that common information can
be shared between VT servers and a VT designer application without
duplicating our IOP parser.
Added missing doxygen.
