+++ isobus/CMakeLists.txt
@@ -50,6 +50,7 @@ set(ISOBUS_SRC
     "isobus_device_descriptor_object_pool_helpers.cpp"
     "can_message_data.cpp"
     "isobus_virtual_terminal_server.cpp"
+    "isobus_virtual_terminal_working_set_base.cpp"
     "isobus_virtual_terminal_server_managed_working_set.cpp")
 
 # Prepend the source directory path to all the source files
@@ -103,6 +104,7 @@ set(ISOBUS_INCLUDE
     "can_message_data.hpp"
     "isobus_virtual_terminal_base.hpp"
     "isobus_virtual_terminal_server.hpp"
+    "isobus_virtual_terminal_working_set_base.hpp"
     "isobus_virtual_terminal_server_managed_working_set.hpp")
 
 # Prepend the include directory path to all the include files
