+++ isobus/include/isobus/isobus/isobus_virtual_terminal_server_managed_working_set.hpp
@@ -17,6 +17,7 @@
 #include "isobus/isobus/can_badge.hpp"
 #include "isobus/isobus/can_control_function.hpp"
 #include "isobus/isobus/isobus_virtual_terminal_objects.hpp"
+#include "isobus/isobus/isobus_virtual_terminal_working_set_base.hpp"
 #include "isobus/utility/event_dispatcher.hpp"
 
 namespace isobus
@@ -26,7 +27,7 @@ namespace isobus
 	/// @brief Defines a managed working set.
 	/// @details This class is meant to be used as the basis for a VT server.
 	/// It keeps track of one active object pool.
-	class VirtualTerminalServerManagedWorkingSet
+	class VirtualTerminalServerManagedWorkingSet : public VirtualTerminalWorkingSetBase
 	{
 	public:
 		/// @brief Enumerates the states of the processing thread for the object pool
@@ -39,8 +40,14 @@ namespace isobus
 			Joined ///< We have sent our response to the working set master and are done parsing
 		};
 
+		/// @brief Default constructor
 		VirtualTerminalServerManagedWorkingSet();
+
+		/// @brief Constructor that takes a control function to associate with this working set
+		/// @param[in] associatedControlFunction The control function to associate with this working set
 		VirtualTerminalServerManagedWorkingSet(std::shared_ptr<ControlFunction> associatedControlFunction);
+
+		/// @brief Destructor
 		~VirtualTerminalServerManagedWorkingSet();
 
 		/// @brief Starts a thread to parse the received object pool files
@@ -49,21 +56,14 @@ namespace isobus
 		/// @brief Joins the parsing thread
 		void join_parsing_thread();
 
-		bool parse_iop_into_objects(std::uint8_t *iopData, std::uint32_t iopLength);
-
 		/// @brief Returns if any object pools are being managed for this working set master
 		/// @returns true if at least 1 object pool has been received for this working set master, otherwise false
 		bool get_any_object_pools() const;
 
-		std::shared_ptr<VTObject> get_object_by_id(std::uint16_t objectID);
-		std::shared_ptr<VTObject> get_working_set_object();
-
+		/// @brief Returns the state of object pool processing, useful when parsing the object pool
+		/// on its own thread.
+		/// @returns The state of object pool processing
 		ObjectPoolProcessingThreadState get_object_pool_processing_state();
-		std::uint16_t get_object_pool_faulting_object_id();
-
-		void add_iop_raw_data(std::vector<std::uint8_t> &dataToAdd);
-		std::size_t get_number_iop_files() const;
-		std::vector<std::uint8_t> &get_iop_raw_data(std::size_t index);
 
 		/// @brief Returns the control function that is the working set master
 		/// @returns The control function that is the working set master
@@ -79,27 +79,19 @@ namespace isobus
 
 		/// @brief Saves an event callback handle for the lifetime of this object
 		/// which is useful for keeping track of callback lifetimes in a VT server
+		/// @param[in] callbackHandle The event callback handle to save
 		void save_callback_handle(isobus::EventCallbackHandle callbackHandle);
 
 		/// @brief Clears all event callback handles for the this working set
 		/// which is useful if you want to stop drawing this working set
 		void clear_callback_handles();
 
-		/// @brief Returns a colour from this working set's current colour table, by index
-		/// @param[in] colourIndex The index into the VT's colour table to retrieve
-		/// @returns A colour from this working set's current colour table, by index
-		VTColourVector get_colour(std::uint8_t colourIndex) const;
-
-		/// @brief Returns the working set's object tree
-		/// @returns The working set's object tree
-		const std::map<std::uint16_t, std::shared_ptr<VTObject>> &get_object_tree() const;
-
 		/// @brief Tells the server where this pool originated from.
 		/// @returns True if this pool was loaded via a Load Version Command, otherwise false (transferred normally)
 		bool get_was_object_pool_loaded_from_non_volatile_memory() const;
 
 		/// @brief Tells the server where this pool originated from.
-		/// @param[in] True if this pool was loaded via a Load Version Command, otherwise false (transferred normally)
+		/// @param[in] value True if this pool was loaded via a Load Version Command, otherwise false (transferred normally)
 		void set_was_object_pool_loaded_from_non_volatile_memory(bool value, CANLibBadge<VirtualTerminalServer>);
 
 		/// @brief Sets the object ID of the currently focused object
@@ -130,54 +122,22 @@ namespace isobus
 		bool is_deletion_requested() const;
 
 	private:
-		/// @brief Adds an object to the object tree, and replaces an object of the same type
-		/// if there's already one in the tree with the same ID.
-		/// @param[in] objectToAdd The object to add to the object tree
-		/// @returns true if the object was added or replaced, otherwise false
-		bool add_or_replace_object(std::shared_ptr<VTObject> objectToAdd);
-
-		/// @brief Parses one object in the remaining object pool data
-		/// @param[in,out] iopData A pointer to some object pool data
-		/// @param[in,out] iopLength The number of bytes remaining in the object pool
-		/// @returns true if an object was parsed
-		bool parse_next_object(std::uint8_t *&iopData, std::uint32_t &iopLength);
-
-		/// @brief Checks if the object pool contains an object with the supplied object ID
-		/// @returns true if an object with the specified ID exists in the object pool
-		bool get_object_id_exists(std::uint16_t objectID);
-
 		/// @brief Sets the object pool processing state to a new value
 		/// @param[in] value The new state of processing the object pool
 		void set_object_pool_processing_state(ObjectPoolProcessingThreadState value);
 
-		/// @brief Sets the object ID associated with a faulting object during pool parsing
-		/// @param[in] value The object ID to set as the faulting object
-		void set_object_pool_faulting_object_id(std::uint16_t value);
-
-		/// @brief Returns the event ID from a byte. Does validation to ensure that the byte is valid.
-		/// If the proprietary range or reserved range is used, it will be considered invalid and event 0 will be returned.
-		/// @param[in] eventByte The byte to convert to an event ID
-		/// @returns The event ID from a byte, or event 0 if the byte is invalid
-		static EventID get_event_from_byte(std::uint8_t eventByte);
-
 		/// @brief The object pool processing thread will execute this function when it runs
 		void worker_thread_function();
 
-		VTColourTable workingSetColourTable; ///< This working set's colour table
-		std::map<std::uint16_t, std::shared_ptr<VTObject>> vtObjectTree; ///< The C++ object representation (deserialized) of the object pool being managed
-		std::vector<std::vector<std::uint8_t>> iopFilesRawData; ///< Raw IOP File data from the client
 		std::unique_ptr<std::thread> objectPoolProcessingThread = nullptr; ///< A thread to process the object pool with, since that can be fairly time consuming.
-		std::mutex managedWorkingSetMutex; ///< A mutex to protect the interface of the managed working set
 		std::shared_ptr<ControlFunction> workingSetControlFunction = nullptr; ///< Stores the control function associated with this working set
 		std::vector<isobus::EventCallbackHandle> callbackHandles; ///< A convenient way to associate callback handles to a working set
 		ObjectPoolProcessingThreadState processingState = ObjectPoolProcessingThreadState::None; ///< Stores the state of processing the object pool
 		std::uint32_t workingSetMaintenanceMessageTimestamp_ms = 0; ///< A timestamp (in ms) to track sending of the maintenance message
 		std::uint32_t auxiliaryInputMaintenanceMessageTimestamp_ms = 0; ///< A timestamp (in ms) to track if/when the working set sent an auxiliary input maintenance message
-		std::uint16_t workingSetID = NULL_OBJECT_ID; ///< Stores the object ID of the working set object itself
-		std::uint16_t faultingObjectID = NULL_OBJECT_ID; ///< Stores the faulting object ID to send to a client when parsing the pool fails
 		std::uint16_t focusedObject = NULL_OBJECT_ID; ///< Stores the object ID of the currently focused object
 		bool wasLoadedFromNonVolatileMemory = false; ///< Used to tell the server how this object pool was obtained
-		bool workingSetDeletionRequested = false;
+		bool workingSetDeletionRequested = false; ///< Used to tell the server to delete this working set
 	};
 } // namespace isobus
